added motors for clawRotation and changed how the claw opens and closes
Defined ClawRotationSlave1 in port1 and ClawRotationSlave2 in port10 and
slaved them to ClawRotation. These motors will be added and wired on
Friday, 1/13/2017.

Changed the way that the claw opened and closed to remove the sleep
methods so that there is no more akward lag when opening and closing the
claw.